convert most weights to bfloat8_b

+++ models/experimental/stable_diffusion3/tt/transformer.py
@@ -49,7 +49,7 @@ class TtSD3Transformer2DModelParameters:
             )
 
         return cls(
-            pos_embed=TtPatchEmbedParameters.from_torch(substate(state, "pos_embed"), dtype=dtype, device=device),
+            pos_embed=TtPatchEmbedParameters.from_torch(substate(state, "pos_embed"), device=device),
             time_text_embed=TtCombinedTimestepTextProjEmbeddingsParameters.from_torch(
                 substate(state, "time_text_embed"), dtype=dtype, device=device
             ),
